Moves EntitlementInfos
